feat(enieber) create test and separate calculateReceivedScore
