+++ utils/index.js
@@ -2,6 +2,7 @@ import config from "config-yml";
 import dotenv from "dotenv";
 import request from "make-requests";
 
+import {  calculateReceivedScore } from './calculateReceivedScore';
 import userController from "../controllers/user";
 import { sendCollect, sendBotCollect } from "./analytics";
 if (process.env.NODE_ENV !== "production") {
@@ -68,29 +69,7 @@ export const calculateScore = interaction => {
   return score;
 };
 
-export const calculateReceivedScore = interaction => {
-  let score = 0;
-  if (interaction.type === "reaction_added") {
-    if (interaction.description === "+1") {
-      score = config.xprules.reactions.receive.positive;
-    } else if (interaction.description === "-1") {
-      score = config.xprules.reactions.receive.negative;
-    } else if (interaction.description === "atena") {
-      score = config.xprules.reactions.atena;
-    }
-  } else if (interaction.type === "reaction_removed") {
-    if (interaction.description === "+1") {
-      score = config.xprules.reactions.receive.positive * -1;
-    } else if (interaction.description === "-1") {
-      score = config.xprules.reactions.receive.negative * -1;
-    } else if (interaction.description === "atena") {
-      score = config.xprules.reactions.atena * -1;
-    }
-  } else if (interaction.type === "thread") {
-    score = config.xprules.threads.receive;
-  }
-  return score;
-};
+export calculateReceivedScore;
 
 export const calculateLevel = score => {
   const level = config.levelrules.levels_range.findIndex(l => score < l) + 1;
